+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/model/BookSummaryModel.kt
@@ -2,7 +2,7 @@ package com.yeobaek.data.model
 
 data class BookSummaryModel(
     val authors: List<String>,
-    val bookId: Int,
+    val bookId: Long,
     val passageCount: Int,
     val title: String,
 )
